backend/src/requirements.txt
aiohttp 
aiosignal
argcomplete==1.10.3
async-timeout
attrs 
awsebcli==3.20.7
bcrypt==4.0.1
beautifulsoup4==4.8.2
bleach 
blessed==1.20.0
botocore==1.29.158
Bottleneck 
brotlipy==0.7.0
cached-property==1.5.2
cement==2.8.2
certifi
cffi 
chardet==3.0.4
charset-normalizer
ci-info==0.3.0
click
colorama==0.4.3
compressed-rtf==1.0.6
configobj==5.0.8
configparser==5.3.0
cryptography
dataclasses-json
defusedxml
dnspython==2.3.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docx2txt==0.8
ebcdic==1.1.1
entrypoints
etelemetry==0.3.0
extract-msg==0.28.7
fastjsonschema
filelock==3.12.2
Flask 
Flask-Cors 
frozenlist
fsspec==2023.6.0
future==0.18.3
greenlet
gunicorn==20.1.0
httplib2==0.22.0
huggingface-hub==0.15.1
idna 
IMAPClient==2.1.0
isodate==0.6.1
itsdangerous 
Jinja2
jmespath==1.0.1
joblib==1.2.0
jsonschema==3.2.0
jupyter_client
jupyter_core
jupyterlab-pygments
langchain
loguru==0.7.0
looseversion==1.2.0
lxml 
MarkupSafe
marshmallow 
marshmallow-enum 
mistune
mpmath==1.3.0
multidict
mypy-extensions==0.4.3
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numexpr 
numpy
olefile==0.46
openai==0.27.8
openapi-schema-pydantic
packaging
pandas==1.5.3
pandocfilters
paramiko==3.2.0
pathlib==1.0.1
pathspec==0.10.1
pdfminer.six==20191110
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs
prov==2.0.0
psycopg2-binary
pycparser
pycryptodome==3.18.0
pydantic
pydot==1.4.2
Pygments
PyMuPDF==1.22.5
PyNaCl==1.5.0
pyOpenSSL
pyparsing==3.1.0
pyrsistent
PySocks
python-dateutil
python-dotenv==1.0.0
python-pptx==0.6.21
pytz
pyxnat==1.5
PyYAML==5.4.1
pyzmq
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.11.0
semantic-version==2.8.5
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.14.0
sortedcontainers==2.4.0
soupsieve
SpeechRecognition==3.8.1
SQLAlchemy
sympy==1.12
tenacity
termcolor==1.1.0
textract==1.6.5
texttable==1.6.7
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado
tqdm
traitlets
traits==6.3.2
transformers==4.30.2
typing-inspect
typing_extensions
tzlocal==5.0.1
urllib3
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl 
